An example of a controller
This is an actual controller from one of my projects, using Zend FW 1 and PHP 5.
The novelty here is the concept of Business Objects - helpers that contain the business logic and are called by the controller to do it's computations. There can be more than one helper per method/command/action, depending on the computation it's performing,  but ideally just one is preferable.

The idea is to rid the controller of business logic, having it's methods be around 5 lines at the most. The resulting code is cleaner, and easier to read and debug.

The business object (BO from now on) will return an object containing the operation's state and messages. It'll evaluate to either true or false in most scenarios, and will contain the pertinent message and data package (usually a JSON object). The method will then be exposed to the frontend, in this case an Ext JS app controller, which will consume the method and apply business logic to deal with the resulting computations and/or errors, if any.