GenerateSubscriptionBO
Part of the setup subsystem that generates dummy data for tests and demo. Calls another BO for specialized computations, and is itself used by the SetUpController.

Fun fact: I posted part of the code on a Stack Overflow tread about single operation SQL inserts. Here's the tread: https://stackoverflow.com/a/29107987/2689491
Got only one up-vote for this particular answer.

Multiple round trips to the database introduce a performance problem and, since you must connect multiple times, a latency problem. The operation takes more time, and has a higher computational cost. With this approach, the script connects to the database only once, and executes a single operation.

The benchmarks on both the multiple connections approach and single connection approach proved that the second option is almost 40% faster. It is also in line with the best practices on the MySQL server documentation.